Now supporting symbolic debugging in AS6.1, compiled via Makefile.
I was also able to compile on OS X (with yagarto-4.7.2), and debug the generated TinyG2.elf file using Atmel Studio 6.1.

I added the Reference directory with a Makefile to gather reference code. This currently only contains newlib-1.20, and is used for symbolic debugging only. To download and prepare all of the reference code, just:

cd Reference
make