Fixed missing ::port definition.
This was tricky. Basically, the pin creates a complete copy of the port object. With optimization OFF, this leaves an actual object and functions floating around that normally wouldn't be, and they get duplicated.

To get around it, I made the Port32 methods I needed static, and used them directly. Arguably, *all* of the Port32 methods could/should be static.