Introducing checksums and sequential-line-number checking.
Use {nxln:0} to reset the next line number expected.

Add *nnn to the end of a line (can be after a ; for gcode cleanness) where non is the decimal (NOT with leading zeroes) value of the XOR of every character before the *.

If there is a *checksum AND a line number, then the checksum is verified and the line number is checked to be the next one it was expecting (the value of nxln or the previous line’s N + 1).