Inch mode precision fixes (#296)
Reports in inches need 3 more decimal places than in mm.
This patch fixes that, while ensuring that every NV object
of TYPE_FLOAT has nv->precision set to a definite value.
The static precision values in the configuration table
are left as-is, dynamically modifying nv->precision in
created objects according to the mm/inches mode.

It also fixes a related problem, in which position changes
sometimes went unreported because of rollover boundary
issues.