Automatic multi-move homing when motors share an axis
IMPORTANT: Requires that moves of all motors of an axis will trigger the same input!

IOW: If motors 2 and 3 are both on the Y axis, you will get three moves:

1. One with BOTH motors until it triggers the input
2. One with motor 2 while motor 3 is "frozen" until the input is triggered
3. One last move with motor 3 while motor 2 is "frozen" until the input is triggered

In order for this to work with two switches, you'll need the input set to `IO_ACTIVE_HIGH` with a physical pull DOWN (to ground), and *both( switches wired with the common side to the input, and the Normally Open (N.O.) side to logic `HIGH` (3.3v or 5v, depending on the board and jumpers, etc.). This way  when either or both switch is tripped it'll set the input to `HIGH`.